templates/test_template.py
"""Test template for API endpoints.

This template should be used for all generated API tests to ensure consistency.
"""
import pytest
from fastapi.testclient import TestClient
from datetime import datetime
from typing import Dict, Any

from app.main import app

class Test{EndpointName}:
    """Test cases for {endpoint_name} endpoint.
    
    Test Coverage:
    - Create operations
    - Read operations
    - Update operations
    - Delete operations
    - Error scenarios
    - Edge cases
    """
    
    @classmethod
    def setup_class(cls):
        """Setup test dependencies at class level."""
        cls.client = TestClient(app)
        cls.base_url = "/api/v1/{endpoint_path}"
        cls.test_data = {}
        
    @classmethod
    def teardown_class(cls):
        """Cleanup after all tests in class."""
        # Clean up any test data if needed
        pass
        
    def setup_method(self, method):
        """Setup before each test method."""
        self.test_id = f"test_{method.__name__}_{datetime.now().timestamp()}"
        
    def teardown_method(self, method):
        """Cleanup after each test method."""
        # Clean up method-specific data
        pass
    
    # ========== CREATE Operations ==========
    
    @pytest.mark.create
    def test_create_{entity}_success(self):
        """Test successful {entity} creation."""
        # Arrange
        payload = {
            "field1": "value1",
            "field2": "value2"
        }
        
        # Act
        response = self.client.post(self.base_url, json=payload)
        
        # Assert
        assert response.status_code == 201
        data = response.json()
        assert data["field1"] == payload["field1"]
        assert "id" in data
        
        # Store for cleanup
        self.__class__.test_data["created_id"] = data["id"]
    
    @pytest.mark.create
    @pytest.mark.parametrize("invalid_payload,expected_error", [
        ({}, "field1 is required"),
        ({"field1": ""}, "field1 cannot be empty"),
        ({"field1": "x" * 256}, "field1 too long"),
    ])
    def test_create_{entity}_validation_errors(self, invalid_payload, expected_error):
        """Test {entity} creation with invalid data."""
        # Act
        response = self.client.post(self.base_url, json=invalid_payload)
        
        # Assert
        assert response.status_code == 422
        assert expected_error in response.json()["detail"]
    
    # ========== READ Operations ==========
    
    @pytest.mark.read
    def test_get_{entity}_by_id_success(self):
        """Test retrieving {entity} by ID."""
        # Arrange
        entity_id = self.test_data.get("created_id", "123")
        
        # Act
        response = self.client.get(f"{self.base_url}/{entity_id}")
        
        # Assert
        assert response.status_code == 200
        data = response.json()
        assert data["id"] == entity_id
    
    @pytest.mark.read
    def test_get_{entity}_not_found(self):
        """Test retrieving non-existent {entity}."""
        # Act
        response = self.client.get(f"{self.base_url}/non-existent-id")
        
        # Assert
        assert response.status_code == 404
        assert "not found" in response.json()["detail"].lower()
    
    @pytest.mark.read
    @pytest.mark.parametrize("filters,expected_count", [
        ({"status": "active"}, 5),
        ({"created_after": "2024-01-01"}, 3),
    ])
    def test_list_{entities}_with_filters(self, filters, expected_count):
        """Test listing {entities} with various filters."""
        # Act
        response = self.client.get(self.base_url, params=filters)
        
        # Assert
        assert response.status_code == 200
        data = response.json()
        assert len(data["items"]) == expected_count
    
    # ========== UPDATE Operations ==========
    
    @pytest.mark.update
    def test_update_{entity}_success(self):
        """Test successful {entity} update."""
        # Arrange
        entity_id = self.test_data.get("created_id", "123")
        update_payload = {
            "field1": "updated_value"
        }
        
        # Act
        response = self.client.patch(
            f"{self.base_url}/{entity_id}", 
            json=update_payload
        )
        
        # Assert
        assert response.status_code == 200
        data = response.json()
        assert data["field1"] == update_payload["field1"]
    
    # ========== DELETE Operations ==========
    
    @pytest.mark.delete
    def test_delete_{entity}_success(self):
        """Test successful {entity} deletion."""
        # Arrange
        entity_id = self.test_data.get("created_id", "123")
        
        # Act
        response = self.client.delete(f"{self.base_url}/{entity_id}")
        
        # Assert
        assert response.status_code == 204
        
        # Verify deletion
        get_response = self.client.get(f"{self.base_url}/{entity_id}")
        assert get_response.status_code == 404
    
    # ========== Error Scenarios ==========
    
    @pytest.mark.error
    def test_{entity}_unauthorized_access(self):
        """Test accessing {entity} without authentication."""
        # Arrange
        headers = {"Authorization": "Bearer invalid-token"}
        
        # Act
        response = self.client.get(self.base_url, headers=headers)
        
        # Assert
        assert response.status_code == 401
    
    # ========== Helper Methods ==========
    
    def _create_test_{entity}(self, **kwargs) -> Dict[str, Any]:
        """Helper method to create test {entity}."""
        default_data = {
            "field1": f"test_{self.test_id}",
            "field2": "default_value"
        }
        default_data.update(kwargs)
        
        response = self.client.post(self.base_url, json=default_data)
        assert response.status_code == 201
        return response.json()
    
    def _assert_valid_{entity}_response(self, data: Dict[str, Any]):
        """Helper to validate {entity} response structure."""
        required_fields = ["id", "field1", "field2", "created_at"]
        for field in required_fields:
            assert field in data, f"Missing required field: {field}"